+++ xocto/types/generic.py
@@ -1,0 +1,21 @@
+from typing import Any, Protocol, TypeVar
+
+
+T = TypeVar("T", covariant=True)
+
+
+class Comparable(Protocol[T]):
+    """
+    Just a very basic way of describing an object that can be compared to another using some
+    basic operations.
+    """
+
+    def __eq__(self, other: Any) -> bool: ...
+
+    def __lt__(self, other: Any) -> bool: ...
+
+    def __le__(self, other: Any) -> bool: ...
+
+    def __gt__(self, other: Any) -> bool: ...
+
+    def __ge__(self, other: Any) -> bool: ...
